Refactor: Command Model 서비스로 이동
